+++ src/lib.rs
@@ -5,6 +5,7 @@
 
 pub mod api;
 pub mod app;
+pub mod cli;
 pub mod config;
 pub mod model;
 pub mod myjd;